Refactor rule files upload for clarity
